feat(event): assign images to events
A new table event_upload is introduced to store the connection between uploaded images and events.
If more than one image is assigned to an event the sorting of these images  will be based on the image file name. To enable this, file_name is now an additional column in the table upload. This table has another new column, file_type, with the default value 'image' as currently only image files can be uploaded.