feat!: add location (#114)
* feat(location): add location data

A new table `location` is created to accommodate location data of an account or an event based on GPS coordinates. The new function `distance' calculates the linear distance between two locations.

* refactor(location): cleanup

* feat(location): make use of PostGIS extension

Instead of working with the original GPS coordinates we use the data type GEOMETRY from the PostGIS extension. Table `maevs.location`was dropped in favor of GEOMETRY typed columns in tables `maevsi_private.account' and `maevsi.event'. A collection of location related functions was added in a single source code file `function_package_location.sql`similar to the package feature in Oracle which is missing in PostgreSQL.

* feat(location): grant execute permissions on postgis functions

Necessary for the Postgraphile PostGIS module.

* feat(location): switch to geography

* chore(location): remove table

* fix(postgis): correct execution grant

* feat(location): add verifications for indexes

Queries on dictionary tables were added to the  verification scripts for spatial indexes.

* refactor: work in feedback

---------

Co-authored-by: Jonas Thelemann <e-mail@jonas-thelemann.de>