feat(friendship): new table friendship_request
There is a new table `friendship_request` holding a single row per friendship request. If the request is accepted then 2 records will be inserted into table `friendship` and the row in `friendship_request` will be deleted. If the request is rejected then the row in `friendship_request`will also be deleted deleted. This makes the enum type `friendship_status` superfluous and also allows policies on table `friendship` in order to hide information about friendship closeness as set by other users.